renamed lib/persistence to lib/Persistence, renamed IEndpointContent interface to IActionsEndpointContent
